+++ GodotTry-on/main_menu.gd
@@ -12,7 +12,7 @@ func _process(delta: float) -> void:
 
 
 func _on_start_pressed() -> void:
-	get_tree().change_scene_to_file("res://EthnicityDetectionScene.tscn")
+	get_tree().change_scene_to_file("res://MaskTryon.tscn")
 
 
 func _on_credit_pressed() -> void: